04_feature_extraction_experiment/requirements.txt
# 基礎音訊處理套件
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.3

# 語音活動檢測套件
webrtcvad>=2.0.10
pyannote.audio>=3.1.1

# 音訊降噪套件
noisereduce>=2.0.0
sounddevice>=0.4.6

# 音量分析套件
pyloudnorm>=0.1.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# 其他工具套件
tqdm>=4.62.0
PyYAML>=5.4.0

# 新增的依賴包
pydub>=0.25.1
scikit-learn>=0.24.0 